+++ start_simulation.sh
@@ -87,7 +87,7 @@ if [[ -n "$SPINE_ARCHIVE" ]] && [[ -z "$BUILD" ]]; then
         # Return code 1 is from closing the simulation GUI
         if [ $SPINE_RC -eq 1 ]; then
             if echo "$OUTPUT" | grep -q "version.*GLIBC"; then
-                echo "It seems your GLIBC version is not compatible with the downloaded binary"
+                echo "⚠️ It seems your GLIBC version is not compatible with the downloaded binary"
                 BUILD=1
             fi
         elif [ $SPINE_RC -ne 0 ] && [ $SPINE_RC -ne 1 ]; then