+++ upkie/envs/wrappers/tests/differentiate_action_test.py
@@ -8,7 +8,7 @@
 
 import unittest
 
-import gymnasium
+import gymnasium as gym
 import numpy as np
 
 from upkie.envs.wrappers.differentiate_action import DifferentiateAction
@@ -31,7 +31,7 @@ class DifferentiateActionTestCase(unittest.TestCase):
         try:
             from stable_baselines3.common.env_checker import check_env
 
-            env = gymnasium.make("Pendulum-v1")
+            env = gym.make("Pendulum-v1")
             diff_env = DifferentiateAction(
                 env,
                 min_derivative=-0.1,