+++ upkie/envs/wrappers/tests/noisify_action_test.py
@@ -8,7 +8,7 @@
 
 import unittest
 
-import gymnasium
+import gymnasium as gym
 import numpy as np
 
 from upkie.envs.wrappers.noisify_action import NoisifyAction
@@ -26,7 +26,7 @@ class NoisifyActionTestCase(unittest.TestCase):
         try:
             from stable_baselines3.common.env_checker import check_env
 
-            env = gymnasium.make("Pendulum-v1")
+            env = gym.make("Pendulum-v1")
             noisy_env = NoisifyAction(env, noise=np.full(1, 0.42))
             check_env(noisy_env)
         except ImportError: